Move NamespaceManager to a new file
